apps/frontend/smile_bot_dupe/requirements.txt
openai
colorama
python-dotenv
streamlit